+++ DarkCodex/changelog.md
@@ -13,6 +13,7 @@
 - option to allow achievements; this also clears the corresponding flag from future save files (otherwise uninstalling all mods would prevent achievements again)
 - impale infusion
 - ice tomb
+- extra rogue talent
 
 ### Changed
 - testing out new parameters on mobile gathering